Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3500



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@15639 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/rbbitst.cpp
@@ -315,7 +315,7 @@ static void printStringBreaks(UnicodeString ustr, int expected[],
             j ++;
         }
         u_charName(c, U_UNICODE_CHAR_NAME, name, 100, &status);
-        printf("%7x %5d %6d %8d %4s %4s %s\n", c, 
+        printf("%7x %5d %6d %8d %4s %4s %s\n", (int)c, 
                            u_isUAlphabetic(c), 
                            u_hasBinaryProperty(c, UCHAR_GRAPHEME_EXTEND),
                            u_isalnum(c),
